Move v4.0.0 specific utils
